Added option to put anchors on lines: 'lineno-anchors', with default value false.
If enabled, adds lineno<linenumber> anchors to all lines.

This is intended to be used when syntaxHighlighting a whole file, or
with a most one source fragment on the page.

It can then be used with something like this:
ref="somefile.html#lineno42".

In the SyntaxHighlighted page, you must write the body.onLoad
handler something this:
onLoad="SyntaxHighlighter.highlight(); location.hash=location.hash;"